work on bugs

+++ backend/controllers/tasks.js
@@ -12,15 +12,17 @@ exports.getTasks = async (req, res, next) => {
     let id = req.params.id
     let newDate = new Date();
     let date = `${transformTime(newDate.getFullYear().toString())}-${transformTime((newDate.getMonth() + 1).toString())}-${transformTime(newDate.getDate().toString())}`
-    try {
 
+
+    try {
+        if (typeof (id) === "undefined") return
         const tasks = await sequelize.query(`
         SELECT firstname, memo, "userId", "Tasks".id, "Tasks"."createdAt", "Tasks".hours, "Tasks".minutes, "Tasks".seconds  FROM "Users"
             JOIN "Tasks"
                 ON "Users".id = "Tasks"."userId"
             WHERE "Users".id = ${id} AND "Tasks"."createdAt" = '${date}'
            `)
-        res.json(tasks[0])
+        res.status(200).json(tasks[0])
 
     } catch (error) {
         console.log(error.message)
@@ -32,14 +34,14 @@ exports.getCompletedTasksForDays = async (req, res, next) => {
         convertedEndDate = req.body.convertedEndDate,
         id = req.body.id;
 
-
     try {
+        if (typeof (id) === "undefined") return
         const tasks = await sequelize.query(`
             SELECT * FROM "Tasks"
             WHERE "userId" = ${id} AND "createdAt" BETWEEN '${convertedStartDate}' AND '${convertedEndDate}'
             ORDER BY "createdAt" ASC;
         `);
-        res.json(tasks[0])
+        res.status(200).json(tasks[0])
 
     } catch (error) {
         console.log(error.message)
@@ -48,15 +50,16 @@ exports.getCompletedTasksForDays = async (req, res, next) => {
 
 
 exports.removeTask = async (req, res, next) => {
-    const id = req.body.id
+    const id = req.body.id;
 
     try {
+        if (typeof (id) === "undefined") return
         await sequelize.query(`
             DELETE FROM "Tasks"
                 WHERE "Tasks".id = '${id}'
         `);
 
-        res.end()
+        res.send(500).end()
 
     } catch (error) {
         console.log(error.message)
@@ -72,6 +75,7 @@ exports.showMatches = async (req, res, next) => {
 
 
     try {
+        if (typeof (id) === "undefined") return
         if (!data) {
             const tasks = await sequelize.query(`
                 SELECT * FROM "Tasks"
@@ -84,7 +88,7 @@ exports.showMatches = async (req, res, next) => {
             WHERE "Tasks".memo LIKE '%${data}%' AND "createdAt" BETWEEN '${convertedStartDate}' AND '${convertedEndDate}' AND "userId" = ${id}
             ORDER BY "createdAt" ASC;
         `);
-        
+
         if (tasks.length === 0) return
         res.status(200).json(tasks[0])
